pyproject.toml
[project]
name = "ai-bagua-oracle"
version = "0.1.0"
description = "AI 八卦骨卜 — I Ching divination backend with bone-casting input and LLM interpretation"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2",
    "anthropic",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
    "ruff",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
